image_server/requirements.txt
# 图片服务器依赖
Pillow>=9.0.0
cryptography>=3.0.0